refactor(agent): move find container to agent cmd
